add conversion back to original format

+++ src/convert_roms.rs
@@ -10,6 +10,7 @@ use rayon::prelude::*;
 use std::env;
 use std::error::Error;
 use std::fs;
+use std::mem::drop;
 use std::path::{Path, PathBuf};
 
 pub fn convert_roms(connection: &PgConnection, matches: &ArgMatches) -> Result<(), Box<dyn Error>> {
@@ -25,6 +26,7 @@ pub fn convert_roms(connection: &PgConnection, matches: &ArgMatches) -> Result<(
             Some("7Z") => to_archive(&connection, &system, &tmp_directory, ArchiveType::SEVENZIP)?,
             Some("CHD") => to_chd(&connection, &system)?,
             Some("CSO") => to_cso(&connection, &system)?,
+            Some("ORIGINAL") => to_original(&connection, &system, &tmp_directory)?,
             Some("ZIP") => to_archive(&connection, &system, &tmp_directory, ArchiveType::ZIP)?,
             Some(_) => bail!("Not implemented"),
             None => bail!("Not possible"),
@@ -83,10 +85,10 @@ fn to_archive(
             });
             add_files_to_archive(&archive_path, &vec![&rom.name], tmp_directory)?;
             fs::remove_file(tmp_directory.join(&rom.name))?;
-            let romfile_input = RomfileInput {
+            let archive_romfile_input = RomfileInput {
                 path: &String::from(archive_path.as_os_str().to_str().unwrap()),
             };
-            update_romfile(connection, &romfile, &romfile_input);
+            update_romfile(connection, &romfile, &archive_romfile_input);
         } else {
             let mut roms: Vec<Rom> = Vec::new();
             let mut romfiles: Vec<Romfile> = Vec::new();
@@ -94,23 +96,29 @@ fn to_archive(
                 roms.push(rom);
                 romfiles.push(romfile);
             }
-            let paths: Vec<&str> = roms.par_iter().map(|rom| rom.name.as_str()).collect();
-            let romfile = romfiles.remove(0);
-            let mut archive_path = Path::new(&romfile.path).to_path_buf();
+            romfiles.dedup();
 
-            extract_files_from_archive(&archive_path, &paths, tmp_directory)?;
+            if romfiles.len() > 1 {
+                bail!("Multiple archives found");
+            }
+
+            let file_names: Vec<&str> = roms.par_iter().map(|rom| rom.name.as_str()).collect();
+            let archive_romfile = romfiles.remove(0);
+            let mut archive_path = Path::new(&archive_romfile.path).to_path_buf();
+
+            extract_files_from_archive(&archive_path, &file_names, tmp_directory)?;
             fs::remove_file(&archive_path)?;
             archive_path.set_extension(match archive_type {
                 ArchiveType::SEVENZIP => "7z",
                 ArchiveType::ZIP => "zip",
             });
-            add_files_to_archive(&archive_path, &paths, tmp_directory)?;
-            for path in paths {
-                fs::remove_file(tmp_directory.join(path))?;
-                let romfile_input = RomfileInput {
+            add_files_to_archive(&archive_path, &file_names, tmp_directory)?;
+            for file_name in file_names {
+                fs::remove_file(tmp_directory.join(file_name))?;
+                let archive_romfile_input = RomfileInput {
                     path: &String::from(archive_path.as_os_str().to_str().unwrap()),
                 };
-                update_romfile(connection, &romfile, &romfile_input);
+                update_romfile(connection, &archive_romfile, &archive_romfile_input);
             }
         }
     }
@@ -128,10 +136,10 @@ fn to_archive(
 
             add_files_to_archive(&archive_path, &vec![&rom.name], &directory)?;
             fs::remove_file(directory.join(&rom.name))?;
-            let romfile_input = RomfileInput {
+            let archive_romfile_input = RomfileInput {
                 path: &String::from(archive_path.as_os_str().to_str().unwrap()),
             };
-            update_romfile(connection, &romfile, &romfile_input);
+            update_romfile(connection, &romfile, &archive_romfile_input);
         } else {
             let mut roms: Vec<Rom> = Vec::new();
             let mut romfiles: Vec<Romfile> = Vec::new();
@@ -139,7 +147,7 @@ fn to_archive(
                 roms.push(rom);
                 romfiles.push(romfile);
             }
-            let paths: Vec<&str> = roms.par_iter().map(|rom| rom.name.as_str()).collect();
+            let file_names: Vec<&str> = roms.par_iter().map(|rom| rom.name.as_str()).collect();
             let directory = Path::new(&romfiles.get(0).unwrap().path)
                 .parent()
                 .unwrap()
@@ -150,16 +158,16 @@ fn to_archive(
                 ArchiveType::ZIP => "zip",
             });
 
-            add_files_to_archive(&archive_path, &paths, &directory)?;
-            for path in paths {
-                fs::remove_file(directory.join(path))?;
+            add_files_to_archive(&archive_path, &file_names, &directory)?;
+            for file_name in file_names {
+                fs::remove_file(directory.join(file_name))?;
             }
-            let romfile_input = RomfileInput {
+            let archive_romfile_input = RomfileInput {
                 path: &String::from(archive_path.as_os_str().to_str().unwrap()),
             };
-            let romfile = create_romfile(connection, &romfile_input);
+            let archive_romfile = create_romfile(connection, &archive_romfile_input);
             for rom in roms {
-                update_rom_romfile(connection, &rom, &romfile.id);
+                update_rom_romfile(connection, &rom, &archive_romfile.id);
             }
             for romfile in romfiles {
                 delete_romfile_by_id(connection, &romfile.id)
@@ -192,10 +200,10 @@ fn to_chd(connection: &PgConnection, system: &System) -> Result<(), Box<dyn Erro
         let (_, cue_romfile) = cue_rom_romfile.remove(0);
         let cue_path = Path::new(&cue_romfile.path).to_path_buf();
         let chd_path = create_chd(&cue_path)?;
-        let romfile_input = RomfileInput {
+        let chd_romfile_input = RomfileInput {
             path: &String::from(chd_path.as_os_str().to_str().unwrap()),
         };
-        let chd_romfile = create_romfile(connection, &romfile_input);
+        let chd_romfile = create_romfile(connection, &chd_romfile_input);
         for (rom, romfile) in bin_roms_romfiles {
             update_rom_romfile(connection, &rom, &chd_romfile.id);
             delete_romfile_by_id(connection, &romfile.id);
@@ -218,13 +226,149 @@ fn to_cso(connection: &PgConnection, system: &System) -> Result<(), Box<dyn Erro
     });
 
     for (_, roms_romfiles) in games_roms_romfiles {
-        for (_, romfile) in roms_romfiles {
-            let iso_path = Path::new(&romfile.path).to_path_buf();
-            let cso_path = create_cso(&iso_path)?;
-            let romfile_input = RomfileInput {
+        for (_, iso_romfile) in roms_romfiles {
+            let iso_path = Path::new(&iso_romfile.path).to_path_buf();
+            let directory = iso_path.parent().unwrap();
+            let cso_path = create_cso(&iso_path, &directory)?;
+            let cso_romfile_input = RomfileInput {
                 path: &String::from(cso_path.as_os_str().to_str().unwrap()),
             };
-            update_romfile(connection, &romfile, &romfile_input);
+            update_romfile(connection, &iso_romfile, &cso_romfile_input);
+        }
+    }
+
+    Ok(())
+}
+
+fn to_original(
+    connection: &PgConnection,
+    system: &System,
+    tmp_directory: &PathBuf,
+) -> Result<(), Box<dyn Error>> {
+    let games_roms_romfiles: Vec<(Game, Vec<(Rom, Romfile)>)> =
+        find_games_roms_romfiles_with_romfile_by_system(connection, &system);
+
+    // partition archives
+    let (archives, others): (
+        Vec<(Game, Vec<(Rom, Romfile)>)>,
+        Vec<(Game, Vec<(Rom, Romfile)>)>,
+    ) = games_roms_romfiles
+        .into_par_iter()
+        .partition(|(_, roms_romfiles)| {
+            roms_romfiles
+                .par_iter()
+                .any(|(_, romfile)| romfile.path.ends_with(".zip") || romfile.path.ends_with(".7z"))
+        });
+
+    // partition CHDs
+    let (chds, others): (
+        Vec<(Game, Vec<(Rom, Romfile)>)>,
+        Vec<(Game, Vec<(Rom, Romfile)>)>,
+    ) = others.into_par_iter().partition(|(_, roms_romfiles)| {
+        roms_romfiles
+            .par_iter()
+            .any(|(_, romfile)| romfile.path.ends_with(".chd"))
+    });
+
+    // partition CSOs
+    let (csos, others): (
+        Vec<(Game, Vec<(Rom, Romfile)>)>,
+        Vec<(Game, Vec<(Rom, Romfile)>)>,
+    ) = others.into_par_iter().partition(|(_, roms_romfiles)| {
+        roms_romfiles
+            .par_iter()
+            .any(|(_, romfile)| romfile.path.ends_with(".cso"))
+    });
+
+    // drop originals
+    drop(others);
+
+    // convert archives
+    for (_, roms_romfiles) in archives {
+        let mut roms: Vec<Rom> = Vec::new();
+        let mut romfiles: Vec<Romfile> = Vec::new();
+        for (rom, romfile) in roms_romfiles {
+            roms.push(rom);
+            romfiles.push(romfile);
+        }
+        romfiles.dedup();
+
+        if romfiles.len() > 1 {
+            bail!("Multiple archives found");
+        }
+
+        let archive_romfile = romfiles.remove(0);
+        let archive_path = Path::new(&archive_romfile.path).to_path_buf();
+        let directory = archive_path.parent().unwrap();
+        let file_names: Vec<&str> = roms.par_iter().map(|rom| rom.name.as_str()).collect();
+        extract_files_from_archive(&archive_path, &file_names, &directory)?;
+        fs::remove_file(&archive_path)?;
+        for rom in roms {
+            let romfile_input = RomfileInput {
+                path: &String::from(directory.join(&rom.name).as_os_str().to_str().unwrap()),
+            };
+            let romfile = create_romfile(connection, &romfile_input);
+            update_rom_romfile(connection, &rom, &romfile.id);
+        }
+        delete_romfile_by_id(connection, &archive_romfile.id);
+    }
+
+    // convert CHDs
+    for (_, roms_romfiles) in chds {
+        let (mut cue_rom_romfile, chd_roms_romfiles): (Vec<(Rom, Romfile)>, Vec<(Rom, Romfile)>) =
+            roms_romfiles
+                .into_par_iter()
+                .partition(|(rom, _)| rom.name.ends_with(".cue"));
+        let (_, cue_romfile) = cue_rom_romfile.remove(0);
+        let cue_path = Path::new(&cue_romfile.path).to_path_buf();
+
+        let mut roms: Vec<Rom> = Vec::new();
+        let mut romfiles: Vec<Romfile> = Vec::new();
+        for (rom, romfile) in chd_roms_romfiles {
+            roms.push(rom);
+            romfiles.push(romfile);
+        }
+        romfiles.dedup();
+
+        if romfiles.len() > 1 {
+            bail!("Multiple CHDs found");
+        }
+
+        let chd_romfile = romfiles.remove(0);
+        let chd_path = Path::new(&chd_romfile.path).to_path_buf();
+        let directory = chd_path.parent().unwrap();
+        let file_names_sizes: Vec<(&str, u64)> = roms
+            .iter()
+            .map(|rom| (rom.name.as_str(), rom.size as u64))
+            .collect();
+        extract_chd(
+            &chd_path,
+            &directory,
+            &tmp_directory,
+            &cue_path.file_name().unwrap().to_str().unwrap(),
+            &file_names_sizes,
+        )?;
+        fs::remove_file(&chd_path)?;
+        for rom in roms {
+            let romfile_input = RomfileInput {
+                path: &String::from(directory.join(&rom.name).as_os_str().to_str().unwrap()),
+            };
+            let romfile = create_romfile(connection, &romfile_input);
+            update_rom_romfile(connection, &rom, &romfile.id);
+        }
+        delete_romfile_by_id(connection, &chd_romfile.id);
+    }
+
+    // convert CSOs
+    for (_, roms_romfiles) in csos {
+        for (_, cso_romfile) in roms_romfiles {
+            let cso_path = Path::new(&cso_romfile.path).to_path_buf();
+            let directory = cso_path.parent().unwrap();
+            let iso_path = extract_cso(&cso_path, &directory)?;
+            let iso_romfile_input = RomfileInput {
+                path: &String::from(iso_path.as_os_str().to_str().unwrap()),
+            };
+            update_romfile(connection, &cso_romfile, &iso_romfile_input);
         }
     }
